requirements.txt
pip>=23
torchvision
lightning
torchmetrics
hydra-core
mlflow
rich
pre-commit
pytest
numpy
pandas
matplotlib
scienceplots
scikit-learn
tqdm
seaborn
ipykernel
optuna
dvc
hydra-optuna-sweeper
optuna-integration[pytorch_lightning]
hydra-colorlog
rootutils
pillow
timm
einops
streamlit
onnx
onnxruntime
onnxscript
